Refactor recording & reporting agent time usage
