+++ services/agents/activities/index.ts
@@ -3,6 +3,5 @@ export { AGENT_TIME_CHARGE_LIMIT_MINUTES } from "./constants";
 export { AgentActivity } from "./deprecated-agent-activity";
 export { hasEnoughAgentTimeCharge } from "./deprecated-has-enough-agent-time-charge";
 export { isAgentTimeAvailable } from "./is-agent-time-available";
-export { recordAgentUsage } from "./record-agent-usage";
 export { saveAgentActivity } from "./save-agent-activity";
 export { getMonthlyBillingCycle } from "./utils";
